backend/requirements.txt
asyncpg
fastapi
pandas
openpyxl>=3.1
psycopg-binary==3.2.9
psycopg2-binary
psycopg[binary]
python-dotenv
sqlalchemy[asyncio]
uvicorn[standard]
itsdangerous
pydantic-settings>=2.0
fastapi-limiter>=0.1.6
redis>=5.0
httpx>=0.27